requirements_model_server.txt
flask==3.0.3
gunicorn==22.0.0
werkzeug==3.0.3

pillow==10.3.0

diffusers==0.27.2
transformers==4.41.2
accelerate==0.30.1
safetensors==0.4.3

python-dotenv==1.0.1

torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu
